refactor: modifica nome do slice
